chore(clerk-js): Move CalloutWithAction under common ui
