+++ packages/clerk-js/src/ui/common/index.ts
@@ -1,5 +1,6 @@
 export * from './BlockButtons';
 export * from './constants';
+export * from './CalloutWithAction';
 export * from './forms';
 export * from './redirects';
 export * from './verification';
